third_party/WebKit/LayoutTests/fast/dom/shadow/content-pseudo-element-overridden.html

<!doctype html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<style>
  .contentClass {
    color: blue;
  }
</style>
</head>
<body class="bodyClass">

  <template id="tmpl">
    <style>
      ::content > * {
        color: red;
      }
    </style>
    <content></content>
  </template>

  <div id="host" class="hostClass">
    <div class="contentClass">
      content
    </div>
  </div>

  <pre id='console'></pre>
</body>
<script>
description('Test for crbug.com/274059. Should be able to override ::content styles in shadow root style sheet from the document.');
var host = document.querySelector('#host');
var root = host.createShadowRoot();
var template = document.querySelector('#tmpl');
root.appendChild(template.content);
shouldBe('window.getComputedStyle(document.querySelector(".contentClass")).color', '"rgb(0, 0, 255)"');
</script>
</html>
